Move some OIDC classes into internal package
