+++ dev/com.ibm.ws.security.openidconnect.client/src/com/ibm/ws/security/openidconnect/client/internal/AccessTokenCacheKey.java
@@ -14,13 +14,13 @@ import java.util.Objects;
 
 import com.ibm.websphere.ras.annotation.Sensitive;
 
-public class AccessTokenCacheKey {
+class AccessTokenCacheKey {
 
     @Sensitive
     private final String accessToken;
     private final String configId;
 
-    public AccessTokenCacheKey(@Sensitive String accessToken, String configId) {
+    AccessTokenCacheKey(@Sensitive String accessToken, String configId) {
         this.accessToken = accessToken;
         this.configId = configId;
     }
