+++ dev/com.ibm.ws.security.openidconnect.client/src/com/ibm/ws/security/openidconnect/client/internal/AccessTokenCacheValue.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 IBM Corporation and others.
+ * Copyright (c) 2021, 2022 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -12,12 +12,12 @@ package com.ibm.ws.security.openidconnect.client.internal;
 
 import com.ibm.ws.webcontainer.security.ProviderAuthenticationResult;
 
-public class AccessTokenCacheValue {
+class AccessTokenCacheValue {
 
     private final String uniqueID;
     private final ProviderAuthenticationResult result;
 
-    public AccessTokenCacheValue(String uniqueID, ProviderAuthenticationResult result) {
+    AccessTokenCacheValue(String uniqueID, ProviderAuthenticationResult result) {
         this.uniqueID = uniqueID;
         this.result = result;
     }
