+++ dev/com.ibm.ws.security.openidconnect.client/src/com/ibm/ws/security/openidconnect/client/internal/OidcClientAuthenticator.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2013, 2021 IBM Corporation and others.
+ * Copyright (c) 2013, 2022 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -9,7 +9,7 @@
  * IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.security.openidconnect.client;
+package com.ibm.ws.security.openidconnect.client.internal;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -164,7 +164,7 @@ public class OidcClientAuthenticator {
      * @param result
      * @return
      */
-    ProviderAuthenticationResult fixSubject(ProviderAuthenticationResult result) {
+    public static ProviderAuthenticationResult fixSubject(ProviderAuthenticationResult result) {
         if (result.getSubject() == null) {
             return result;
         }
